Moved cookies_policy_helper to Web engine and respective spec
